backend/requirements.txt
fastapi
uvicorn[standard]
matplotlib
pydantic
python-dotenv
livekit-agents
livekit-server-sdk
livekit-plugins-silero
livekit-plugins-noise-cancellation~=0.2
livekit-agents[openai]~=1.0
PyJWT
python-multipart